backend/requirements.txt
python-telegram-bot>=21.0
pymongo>=4.6.0
motor>=3.3.0
fastapi>=0.115.0
uvicorn>=0.24.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
structlog>=24.1.0
slowapi>=0.1.9
httpx>=0.25.0
requests>=2.31.0
nest_asyncio>=1.5.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
